chore: flytt util til egen fil
